requirements.txt
adium-theme-ubuntu==0.3.4
asn1crypto==0.22.0
attrs==17.2.0
Automat==0.6.0
beautifulsoup4==4.4.1
boto==2.38.0
certifi==2017.4.17
cffi==1.10.0
chardet==3.0.4
constantly==15.1.0
cryptography==1.9
cssselect==1.0.1
decorator==4.0.6
Django==1.8.7
enum34==1.1.6
googlemaps==2.5.1
gspread==0.6.2
html5lib==0.999
httplib2==0.10.3
hyperlink==17.2.1
idna==2.5
incremental==17.5.0
ipaddress==1.0.18
ipython==2.4.1
lxml==3.8.0
mysqlclient==1.3.7
ndg-httpsclient==0.4.0
oauth2client==4.1.2
PAM==0.4.2
parsel==1.2.0
pexpect==4.0.1
Pillow==3.1.2
ptyprocess==0.5
pyasn1==0.2.3
pyasn1-modules==0.0.9
pycparser==2.17
pycrypto==2.6.1
PyDispatcher==2.0.5
Pygments==2.1
PyJWT==1.5.0
pyodbc==4.0.16
pyOpenSSL==17.0.0
pyserial==3.0.1
python-dateutil==2.6.1
pytz==2017.2
queuelib==1.4.2
requests==2.17.3
rsa==3.4.2
Scrapy==1.4.0
selenium==3.4.3
service-identity==17.0.0
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.10
sqlparse==0.1.18
twilio==6.4.1
Twisted==17.5.0
unity-lens-photos==1.0
urllib3==1.21.1
w3lib==1.17.0
zope.interface==4.4.2
